feat: 로그인 세션 연동 + 예약 승인제 전환·취소 모달·마감 캘린더·만나서결제 추가
- 로그인 세션(LoginMemberDto)에서 회원 식별하도록 연동, memberId 타입 Integer로 통일
- 결제완료→관리자 승인→예약확정→이용완료로 상태 흐름 변경(CONFIRMED 추가)
- 예약취소 사유 입력을 alert→모달(select+기타 직접입력)로 변경
- 예약 날짜를 캘린더로 교체, 숙박은 이미 예약된 날 마감(빨강) 처리(서버 이중 체크)
- 관광지·맛집은 예약금만 결제하는 만나서결제 흐름 추가(place_type 컬럼)
- 예약 관련 화면 반응형 대응

+++ src/main/java/com/gnagnoohc/travel/reservation/controller/PaymentController.java
@@ -87,7 +87,7 @@ public class PaymentController {
         String tid = (String) session.getAttribute("KAKAO_TID");
         String orderId = (String) session.getAttribute("KAKAO_ORDER_ID");
         Integer amount = (Integer) session.getAttribute("KAKAO_AMOUNT");
-        Long memberId = (Long) session.getAttribute("KAKAO_MEMBER_ID");
+        Integer memberId = (Integer) session.getAttribute("KAKAO_MEMBER_ID");
 
         log.info("[카카오 success 콜백] reservationId={}, 세션값 tid={}, orderId={}, amount={}",
                 reservationId, tid, orderId, amount);
